requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pyrogram==2.0.106
tgcrypto==1.2.5
httpx==0.27.0
supabase==2.4.0
google-generativeai==0.5.4
cloudinary==1.40.0
python-dotenv==1.0.1
pydantic==2.7.1
pillow==10.3.0
PyJWT==2.8.0
resend==0.7.2
